1.7.4
Updated jQuery window width to inner width (with IE fallback) as it was
giving inconsistent results with/without scrollbar. Added a scroll bar
for IE7 & 8 as they were getting confused when content was expanding
and enabling the scrollbar, triggering the resize event and killing the
menu again. New minified version too.